placeholder tests refactored
