+++ misc/ui/src/test/integration/org/richfaces/component/placeholder/TestPlaceholderCalendar.java
@@ -38,21 +38,6 @@ public class TestPlaceholderCalendar extends AbstractPlaceholderTest {
     private Input secondInput;
 
     @Override
-    String testedComponent() {
-        return "calendar";
-    }
-
-    @Override
-    protected String getTestedValueResponse() {
-        return "Wed Dec 12 00:00:00 CET 2012";
-    }
-
-    @Override
-    protected String getTestedValue() {
-        return "Dec 12, 2012";
-    }
-
-    @Override
     protected Color getDefaultInputColor() {
         return new Color(26, 26, 26);
     }
@@ -62,78 +47,29 @@ public class TestPlaceholderCalendar extends AbstractPlaceholderTest {
         return firstInput;
     }
 
+    @Override
     public Input getSecondInput() {
         return secondInput;
     }
 
-    @Test
-    public void testConverter() {
-        super.testConverter();
+    @Override
+    protected String getTestedValue() {
+        return "Dec 12, 2012";
     }
 
-    @Test
-    public void testDefaultAttributes() {
-        super.testDefaultAttributes();
-    }
-
-    @Test
-    public void testRendered() {
-        super.testRendered();
-    }
-
-    @Test
-    public void testSelector() {
-        super.testSelector();
-    }
-
-    @Test
-    public void testSelectorEmpty() {
-        super.testSelectorEmpty();
-    }
-
-    @Test
-    public void testStyleClass() {
-        super.testStyleClass();
-    }
-
-    @Test
-    public void when_input_with_placeholder_gains_focus_then_placeholder_is_removed() {
-        super.when_input_with_placeholder_gains_focus_then_placeholder_is_removed();
-    }
-
-    @Test
-    public void when_text_is_changed_then_text_changes_color_to_default_and_removes_placeholder_style_classes() {
-        super.when_text_is_changed_then_text_changes_color_to_default_and_removes_placeholder_style_classes();
-    }
-
-    @Test
-    public void when_text_is_cleared_then_input_gets_placeholder_text_and_style_again() {
-        super.when_text_is_cleared_then_input_gets_placeholder_text_and_style_again();
-    }
-
-    @Test
-    public void when_text_is_changed_and_input_is_blurred_then_typed_text_is_preserved() {
-        super.when_text_is_changed_and_input_is_blurred_then_typed_text_is_preserved();
-    }
-
-    @Test
-    public void testAjaxSendsEmptyValue() {
-        super.testAjaxSendsEmptyValue();
-    }
-
-    @Test
-    public void testAjaxSendsTextValue() {
-        super.testAjaxSendsTextValue();
-    }
-
-    @Test
-    public void testSubmitEmptyValue() {
-        super.testSubmitEmptyValue();
+    @Override
+    protected String getTestedValueResponse() {
+        return "Wed Dec 12 00:00:00 CET 2012";
     }
 
     @Ignore("calendar date conversion problem")
     @Test
     public void testSubmitTextValue() {
         super.testSubmitTextValue();
+    }
+
+    @Override
+    String testedComponent() {
+        return "calendar";
     }
 }
